starlight: [ucsb] remove docs site
we're now planning to roll out a new staging system and repurpose the existing
one as the docs site.

this will present some downsides, mainly that databases and k8s namespaces will
be unusual (at least in the short term), but it does allow us to get to the
ideal infrastructure more quickly.

start by dropping the `surfliner-docs` deployment, and add ingress via
docs.digital.library.ucsb.edu to the existing staging site. next i'll introduce
a new staging site in a new namespace and with a new database.