comet: Override Hydra Editor™ edit field partial
With this commit, properties defined in the schema will use a “schema
field” partial and SimpleForm input type that we control. At present
this is mostly identical to prior behaviour, although it does now
use the property definitions defined in the M3.

All properties are multi‐valued for now because they are stored that way
in the database. We will need to do further work to get them to display
as single‐valued in the form, although `definition.multiple?` can be
easily used to check for this.