surfliner_schema: Allow filtering of division props
This commit adjusts `SurflinerSchema::Loader#class_division_for` to
optionally take a block. When it does, the block is used as a filter
for the properties in the resulting `SurflinerSchema::Division`.

This functionality is necessary to allow for the creation of divisions
which contain only primary or nonprimary properties in ~comet.