comet: Add method to test schemaness of forms
Because many form‐rendering views are generic across form types, it’s
useful to be able to check whether a given form has included a
`SurflinerSchema::FormFields` module or not. This commit adds that
check to `Hyrax::Forms::ResourceForm` as `:schema_derived?`. Right now
it is true for PCDM object forms but false for other types of form.

This method still doesn’t exist on forms which aren’t derived from
`Hyrax::Forms::ResourceForm` (collection forms, notably), so
unfortunately it’s still necessary to check for method presence before
calling it.