comet: document some of Comet's fixity goals
this proposed documentation is intended to provide an initial articulation of
Comet's fixity needs on ingest, from which we can open a ticket to check the
implementation.

other parts of the fixity suite need continued discussion, and the goal of
getting this up is to provide an artifact for that discussion.