surfliner_schema: Ignore unavailable properties
If a property has a null or empty `available_on`, `surfliner_schema`
shouldn’t try to process it. The intent here is to enable “abstract”
properties like the following :—

```yaml
properties:
  - &abstract_base_property
    name: null # this would probably cause problems if processed
    available_on: null # this is an abstract base property
    mapping:
      my_mapping: example:ontology/BaseProperty

  - &derived_property
    <<: *abstract_base_property
    name: derived_property
    available_on:
      - generic_object
    # inherits `mapping` from &abstract_base_property
```

Irrespective of the above, processing properties without an availability
is pointless because they are inaccessible in the resulting reader.