comet: Add a smaller M3 schema for testing/demos
This file is aimed at providing a minimal example of what kinds of
things we support in Comet, with pretty heavy documentation in the
form of comments.

One big immediate advantage to a separate file like this is it
means we can demo things like metadata sectioning/groupings prior
to changes being made in the institution‐specific M3 schemas (and it
generally gives developers a place to experiment without worrying about
messing with schemas which are actually in use).