surfliner_schema: Support mapping metadata
This commit adds support for accessing various bits of metadata attached
to mappings, like their `display_label` and so forth. We don’t have any
immediate need for this, since mappings are currently only surfaced
through ~superskunk (which has no use for human‐readable names), but I
needed it for use in the demo with Samvera Metadata Interest Group.