surfliner_schema: Simplify mappings object
The format of the mappings object was made more complicated during the
initial implementation of GeoMetadata mappings, because we initially
derived the mapping values from the property names in the schema. The
current implementation no longer does this, so this reverts the change
in e757f81f489d6bdf37589efa76f842190a0a0b05 and makes the mappings
object a simple `property_iri => value` hash again.

This makes the current common use‐case of “getting the value of a given
property in the mapping” much easier at the cost of no longer being able
to tell from which original propert(y/ies) in the model those values
were derived.