comet: reduce browser interactions in add_object_to_collection_spec
these tests relied on browser interactions extensively for setup. since these
can be complicated and often give rise to intermittent/docker service errors,
especially in the CI environment, it's a good idea to use other tools to setup
the test. use the browser only for the parts of the spec that are intended to
test the full stack.