comet: Use normal test metadata, not a fixture
Previously, comet used the metadata file at
`spec/fixtures/metadata/m3-metadata.yml` to run its tests. This switches
to using the metadata file at `config/metadata/test-metadata.yaml`
instead. The latter file is more accessible and better‐documented.

In general, I tried to keep actual spec changes to a minimum. I added a
`creator` field to the test metadata as some specs use it. Some of the
tests I wrote for M3 behaviours tested a `language` field, but testing
this was unnecessary so I removed it from the spec. The test for M3
property loading needed to be rewritten as the structure of the
`test-metadata` YAML file differs from the old `m3-metadata` one.