comet: Actually test date indexing
A test for this was written a while ago, but was ignored because
the ~surfliner_schema gem didn’t support datatyping properly yet. It has
for a while, so we can turn this test back on.

Right now this tests a `date_created` property with a datatype of
`xsd:date`. We would like to be testing `edtf:EDTF`, but this doesn’t
work yet (it requires the `rdf-edtf` gem).

(This commit also makes a small fix to the GenericObject spec for
datatype support there as well.)