comet: Improve indexing of RDF literals
The `object` of RDF literals is only meaningfully set in subclasses;
otherwise it just takes the value of whatever was the first thing passed
to the initializer. This may be another RDF literal, or some other
object with an undesirable `to_json`. In these cases, we want to use the
literal (string) `value` when indexing instead.

This fixes indexing issues when a property has an unrecognized RDF
datatype, and probably some other edge cases as well.