surfliner_schema: Fix casting of RDF literals
Formerly, the gem exhibited weird behaviour when setting a date(time)
property to an invalid value: Instead of preserving the (invalid)
lexical value, it would be replaced with the current date. This is
because the `RDF::Literal::Date` constructor only remembers the lexical
value if its first argument is a string (or if the lexical value is
provided explicitly with the `lexical:` kwarg), and it uses the current
date as a fallback when the provided date is invalid.

We want to preserve lexical values on coersion, not replace them, so
this commit adds an extra test for this case.