starlight: update to Ruby 3.1.3
The previous update didn't actually update the ruby version, since the
`RUBY_VERSION` `ARG` in the ~starlight `Dockerfile` was not actually
used in any way during the build process, because the `base` image being
relied on is built completely independently.

This also updated the `RUBY_ABI` to `3.1.0` to match the new major/minor
version of Ruby.

Note at this time we cannot update to Ruby 3.2 as our current version of
Rails (6.x) is not compatible with Ruby 3.2