docker: support specifying alpine version
We are in a situation where we still have a few apps running Ruby 2.7,
which does not have an Alpine 3.17-based docker image available. We
should support building base images where we can specify both the ruby
and corresponding alpine version for the image build so we're not stuck
to a single version of alpine.