comet: Use test‐metadata in local dev environment
This makes the development environment in `docker-compose` use the same
schema as the test environment. Apparently tests in `docker-compose`
currently don’t properly load the test environment and just run in
`development` instead, which we should probably fix, but this isn’t a
bad change to make in the meantime.