comet: document initial visibility model` decisions
as part of #977, we settled on a cross-campus visibility model as what is
hopefully a minor customization of Hyrax's built-in model. this documents the
basic structure and agreed upon conditions so this rule set can be implemented.

another round of documentation is needed here to provide some developer guidance
about the use of the `#visibility` accessor and its underlying tools. and
perhaps more to document the desired semantics of these visibility levels and
the proposed group membership rules implied.