comet: update i18n values per #1255
some of the changes included in this commit depend on an update to the bulkrax
version which can't yet be merged due to a possible bug in the bulkrax `main`
branch.

it's okay to just add the i18n keys in the meanwhile, though, and this way we
can merge the whole set of label changes together, keeping the dependency update
separate.

related to !3106.

see: https://github.com/samvera-labs/bulkrax/pull/783