comet: allow multi-value columns to import
Columns using a delimiter, with multiple values, should import
successfully for properties that support multiple values.

At the moment, bulkrax treats splitting column values as a completely
separate concern, distinct from whether the column itself supports
multiple values. Since we are using m3/dynamic metadata, we should be
able to directly infer that if a property is defined as multiple, we
can/should split a string value that uses one of the supported
delimiter characters.