comet: track the latest Bulkrax releases
~~we should track `main` on our `trunk` since we're merging things there we'll
need.~~

`bulkrax` is getting really frequent releases. since this is true, and we can
probably influence that pattern continuing, there's a major benefit to pinning
to a release. our bots should help us keep up to date.

if we need to move back to the `hyrax-4-rails-6` branch to get support for those
dependencies, we can always rebase it on `main`. hopefully that should be a
short lived situtation. the bulkrax team is responsive to PRs there and will be
eager to get support in.