comet: split/join on | in bulkrax csv
It seems that it would be desirable to support splitting on as few
delimiter tokens as possible. Rather than supporting `:`, `;` and `|`,
this commit proposes limiting the options to just a `|` pipe character,
with an optional space on either side.

For export, we support a single character for `join`-ing multi-value
elements anyway. So it seems like the least complexity to set both as a
pipe.