surfliner_schema: Support controlled_values
Adds support for a `controlled_values` property on properties, with
`source` and `values` subproperties. `values` is a ~surfliner_schema
extension. No support is given for `format` because we don’t have a good
use‐case for it (yet); it shouldn’t be hard to add if it is needed.

The `values` property mirrors a lot of other ~surfliner_schema
properties in taking either an array or a hash; this is probably
unnecessary but easier than trying to enforce stricter requirements on
what shape it should take. As elsewhere, if a value is not given a
`name` it will be assigned an underscored numeric one; fallback IRIs in
the `about:surfliner_schema/controlled_values` namespace will also be
created if none is provided. Consequently, a minimally‐useful controlled
value might only have a `display_label`. However, giving values a `name`
is good practice as it wards against value reordering in the future.