surfliner_schema: Add support for rdf:langString
Changes the default datatype from `xsd:string` to `rdf:PlainLiteral`.
The latter is treated as functionally equivalent (it also produces
`xsd:string` literals), except that when provided with a
`rdf:langString` value, it is preserved (instead of cast to
`xsd:string`, losing the language information).

(~surfliner_schema will never produce literals with an actual datatype
of `rdf:PlainLiteral`. Literals will always be either `rdf:langString`
or `xsd:string`.)

A type of `rdf:langString` is now supported and handled specially, for
*mandating* that every value have a language tag. `und` (undetermined)
will be used if none is provided.