surfliner_schema: Supported ranges in Houndstooth
The `supported_ranges` method was removed from
`SurflinerSchema::Reader::Base` in
a283039a2735d8c6b289e6f23591e6fdbcf30020. This is generally fine, but we
actually used it in `SurflinerSchema::Reader::Houndstooth` to skip
properties with ranges that were unknown.

This hasn’t caused us any problems so far because we don’t currently use
properties with ranges!

This commit adds it back in as a variable in the reader initializer,
providing a list of all the defined class URIs.