superskunk: use METADATA_DATABASE_NAME everywhere
We have had a mix of database variables for superskunk:

- `METADATA_DATABASE_NAME`
- `POSTGRES_DB`
- `DB_NAME`
- `POSTGRESQL_DATABASE`

And the reality is that superskunk only needs to connect to a single
database. Either a local, standalone DB for local development and
running tests, or a pointing to the pre-existing ~comet metadata
database.

It seemed challenging to understand where a given variable is used, and
why. For example in the `scripts/db-migrate-seed.sh` file, the `DB_NAME`
variable was exported but never really used, whereas the
`METADATA_DATABASE_NAME` variable was used for connection testing and
creation (if necessary). The `okcomputer` health check, however, tested
against the `POSTGRES_DB` variable. So if that variable wasn't in sync
with `METADATA_DATABASE_NAME` a deployment could easily be doing an
irrelevant health check.

The proposal here is to standardize on `METADATA_DATABASE_NAME`, which
seems like it was setup intentionally to make it clear that ~superskunk
is meant to connect to a database full of metadata records.