superskunk: use POSTGRESQL_* variables everywhere
We are also removing the redundant `DB_*` variables. This naming
convention was likely ported from ~comet, which makes sense in that
context as ~comet has both a metadata db (DB_) and internal database
(POSTGRESQL_) to manage. ~superskunk however, only has one database. So
we were maintaining two sets of the same variables.