surfliner_schema: Move class resolution into Reader
This is primarily to remove the circular dependency in `#dry_range`,
which previously called out to the Valkyrie class resolver (which
ideally would be set up to reference a `SurflinerSchema::Loader` which
had the same `Reader`). It’s a lot simpler now, but with the side‐effect
that range properties will now *always* instantiate the class defined in
the same schema file (even if a different class by that name has been
defined as the global constant). It’s not possible (anymore) to
reference classes in Ruby (not defined in the M3) from the M3 file.

I think this is fine and maybe even desirable behaviour. It would be
weird for a class to have a range defined in the schema but for that to
use a different class in actuality because a different constant had been
defined. *However* this does mean that *all* ranges need to go through
the schema class resolver, not be a manually‐declared class which just
happens to include `SurflinerSchema::Schema`. This is again probably a
good pattern to enforce anyway, as the resolver adds a few other
important things as well (namely `.availability`).

There is still a circular dependency in `SurflinerSchema::Mappings`
(which currently depends on `Loader`); there is an open TODO to extract
this behaviour out into a helper service.