surfliner_schema: Remember reader; simplify mapping
Since the `#mapped_to` method was being supplied via the schema class
resolver anyway, the simplest solution was simply to remember the
`SurflinerSchema::Reader` which was used to construct the class and
then call back to it when getting the mappings. We can probably use
this to simplify some other parts of the codebase which previously
relied on an indirection through `SurflinerSchema::Loader` as well.