surfliner_schema: Support profile
We want to use the `additional_metadata` hash here (an extension) to
keep track of the OpenGeoMetadata provider name for ~superskunk. We can
always add additional fields later.

Assuming a class was created through the ~surfliner_schema resource
class resolver, you should be able to get the profile of an object with

```rb
model_class.reader.profile
```