comet: install OTel instrumetation
install the Ruby OpenTelemetry SDK/API, auto instrumentation, and otlp
exporter (to export telemetry data to the otlp based Collector).

for now, install `opentelemetry-instrumentation-all` instead of the specific
instrumentation gems we'll use. we probably want to change this to use the gems
we know we'll want more specifically, but this way we can get a sense for which
of these are activating and which are producing useful data