surfliner_schema: Support “value wrapper” objects
The intent here is to enable a special class of nested Valkyrie objects
which can be easily treated as though they were literals in many cases.
The immediate use‐case is for dates with associated labels, where most
of the time we just want to be dealing with the machine date, but a
human‐readable label is required for display purposes. These objects can
be modelled as having a single `rdf:value` property (providing the
value being wrapped) and any number of additional properties providing
further metadata.

After this change, the following should work regardless of whether
`object.date` is a literal or a nested object with an `rdf:value` :—

```rb
object.date = "1972-12-31"
```

Similarly, the following should cast all dates to literals regardless
of how they are stored :—

```rb
object.date.flat_map(&:terms)
```

This should allow us to continue treating dates “simply” for ingest
and indexing except in cases where the additional complexity is
required.