comet: place all bulkrax files in one FileSet
This is assumed to be a temporary measure that will be more properly
addressed in [1].

For now, however, we satisfy the ~shoreline use case by assigning all
files referenced in bulkrax for a given row/CSVEntry to a single
FileSet. Upon re-ingest, a new FileSet will be created (again, for now).

1. https://gitlab.com/surfliner/surfliner/-/issues/1317