superskunk: allow toggling comet/standalone db
There are some review app situations where we will want to ensure there
is a ~comet deployed for integration testing. However, we would also
like to sometimes have an isolated ~superskunk deployment which does not
need or rely on a ~comet deployment.

This solution uses the gitlab `rules:variables` [1] option as a solution
to this. It does result in some duplicate `HELM_UPGRADE_EXTRA_ARGS`.

1. https://docs.gitlab.com/ee/ci/yaml/#rulesvariables