WIP: add a (possibly) failing test for bulkrax collection creation
we think our hard-coded allow of `admin_set_id` is an issue for Collection
creation.

we'll have CSVs with this field (because objects are included in the CSV), but
Collections don't support this field at all. Normally, we filter fields based on
model, but this one we've hard coded an allow which should only apply to
pcdm:Object classes.