superskunk: align bounding box m3 property names
It is preferred to use the `_geo` suffixed m3 property names for
aardvark/shoreline fields. So we need to update the code that generates
the `dcat_bbox` value to use those names. And the test schema
accordingly. The UCSD and UCSB schemas already have these defined, so no
changes should be needed there.