comet: log missing bucket keys in bulkrax upload
We have already encountered a situation where a user is referencing
staging area files(keys) in a staging area bucket for upload into ~comet
that don't actually exist in that bucket. This attempts to surface that
in a way that's more straightforward to detect in logs.

At the moment when a key doesn't exist in fog, it returns `nil` [1]
rather than raise an error.

1. https://github.com/fog/fog-aws/blob/v3.19.0/lib/fog/aws/models/storage/files.rb#L80