surfliner_schema: Refactor Reform FormFields
It’s a lot easier (and more useful) to generate Reform form fields
directly from the schema reader than to route it through the loader. We
keep track of the reader that was used to generate model classes now, so
this is easy to do if we just make `SurflinerSchema::FormFields` take
the base model class as its argument, instead of a loader and an
availability.

This requires moving the code for dividing up properties into divisions
onto the reader itself rather than the loader, but I think this makes
more sense anyway.

There shouldn’t be any noticeable effects from this commit, but I think
it is a prerequisite for more complex Reform behaviours (e·g references
to other classes in the reader, in the case of nested objects).