comet: setup initial tests for `CometCsvParser`
this code wants more unit tests!

we'll need them to be able to test more complex behavior like binding
relationships between PCDM classes in different configurations. it won't be
practical to do this with browser drivent tests, and the `ValkyrieObjectFactory`
only works with single records.

it seemed like the reason this wasn't well tested is that the setup is a bit
complicated, so let's knock that out.