surfliner_schema: Use Dry::Validation in forms
Minimally, this commit overrides `#deserialize!` in forms which include
`SurflinerSchema::FormFields` to pass the params object through a
`Dry::Validation::Contract` *prior* to deserializing into the form. It
then hacks into the `@result` instance property (defined by Reform) to
record any errors from this process and returns the result as the new
params object Reform will be handling. No actual validations are defined
in Reform proper, as we have determined that `Dry::Validation` is a
better API for doing this sort of work.

We could use `to_result` instead of `@result` to access the Reform result
object, but comments in the Reform source code make clear that the former
should not be considered a stable API and leveraging the internal
instance property seemed slightly less likely to break in the future.

A number of other changes are made to enable the above:

- `SurflinerSchema::FormDefinition` no longer provides `form_options` and
  instead provides `schema_constraints`. Cardinality is now validated for
  all properties thanks to the latter.

- There is a new `.property_transform_for` hook on
  `SurflinerSchema::Loader`s which can be used to apply additional
  transforms to values during the schema coercion step; the intent is
  that this can be used for controlled value coercion if needed.

- `SurflinerSchema::Reader` now provides a `#dry_schema` method which
  can be used in ~lark to define its contracts as well.