comet: include the ark in the published payload
Applications consuming comet publish messages may choose to use the ark,
rather than the comet identifier, as the primary identifier for
resources. An unpublish message needs to contain the ark so that those
applications can have an identifier that is actionable.

If an ark is not minted for the resource at the time of publishing, it
is not included in the payload.