shoreline: support unpublishing arks
We need to support both the case where an ARK is used as the identifier
for an object in shoreline as well as when the resource id is used. We
can now expect an ark in the payload from comet when arks are in use,
and we use the formatted version for unpublish/delete events.